Restructure classes architecture.
