+++ src/services/identity/postgres.ts
@@ -32,8 +32,8 @@ import { Connection } from '../../database/connection/connection.js'
 import type { CustomerEntity } from '../../database/entities/customer.entity.js'
 import { CustomerService } from '../customer.js'
 import { Veramo } from './agent.js'
-import { DefaultIdentity } from './identity.js'
 import type { AbstractPrivateKeyStore } from '@veramo/key-manager'
+import { DefaultIdentityService } from './default.js'
 
 dotenv.config()
 
@@ -43,7 +43,7 @@ const {
 	EXTERNAL_DB_ENCRYPTION_KEY,
 } = process.env
 
-export class PostgresIdentity extends DefaultIdentity {
+export class PostgresIdentity extends DefaultIdentityService {
 	privateStore?: AbstractPrivateKeyStore
 
 	initAgent() {
